refactor(ui): address Greptile review findings on PR #536
Apply two P2 (style-level) improvements from Greptile gatekeeper review:

1. Consistent snapshot pattern in flushPendingDeltasForMessage
   - Collect keys first, then iterate to flush (matches clearPendingDeltasForPart)
   - Prevents mutation-during-iteration and ensures delete() runs before apply
   - If applyPartDeltaV2 throws, entry is already removed (no re-apply on timer)

2. Extract delta-buffer to dedicated module
   - session-events.ts: 854 → 808 lines (under 850, closer to 800 target)
   - delta-buffer.ts: 80 lines (focused, single-responsibility)
   - Improves maintainability per AGENTS.md file-length guidelines

No functional changes. Build verified. Addresses:
https://github.com/Pagecran/CodeNomad/pull/17#discussion_r3383913782
https://github.com/Pagecran/CodeNomad/pull/17#discussion_r3383913841
